Commited Changes 6/22/2019
I have updated the body of the page where I have added elements:  hovering ability for the images.  Linking them to click and go to another page for individual images.  Css elements for creating a grayscale when hovered over.  And and clicking the images goes to the larger format of the picture.